Extract utility functions for querying annotation metadata into a separate module
